Consider child_payloads' keys as well as the parent's
If get_pings_properties is called on a keyed histogram with with_processes=True
then we should include all keys that are present in either the parent or one
of the children.

Otherwise, for a keyed histogram K_HIST with only one key, aKey, only present
in the child, no result will be returned.

Fixes #50